implementado Skills::Loops::Circle que faz o robo andar em circulo com centro, raio e velocidade arbitrarios, o sinal do raio define o sentido (positivo:anti-horario), o sinal da velocidade não importa.
git-svn-id: http://roboime.com.br/svn/RoboCup/Projeto Computacional/RobotIME/trunk@224 536dbed5-76cd-454d-90a9-4f62b6fdd354

+++ Inteligencia/CommanderSIM.cpp
@@ -24,7 +24,7 @@ namespace Inteligencia {
 	CommanderSIM::~CommanderSIM() {}
 
 	void CommanderSIM::prepare() {
-		for(unsigned int n=_robot.size(); n>0; n--) {
+		for(size_t n=_robot.size(); n>0; n--) {
 			stringstream out;
 			Robot* r = _robot[n-1];
 			Command* c = r->command();
